+++ internal/provider/models.go
@@ -96,6 +96,7 @@ type CockroachCluster struct {
 	ServerlessConfig *ServerlessClusterConfig `tfsdk:"serverless"`
 	Regions          []Region                 `tfsdk:"regions"`
 	CockroachVersion types.String             `tfsdk:"cockroach_version"`
+	FullVersion      types.String             `tfsdk:"full_version"`
 	Plan             types.String             `tfsdk:"plan"`
 	State            types.String             `tfsdk:"state"`
 	CreatorId        types.String             `tfsdk:"creator_id"`