+++ CHANGELOG.md
@@ -5,6 +5,11 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+### Added
+
+- Added `full_version` attribute to the cluster data source and resource for
+  fetching the full version string. (e.g. v25.1.0)
+
 ## [1.11.2] - 2025-02-07
 
 ### Fixed